refactor: изменил структуру проекта
