+++ src/main/java/org/example/tonpad/core/parser/extension/block/AbstractBlock.java
@@ -1,0 +1,8 @@
+package org.example.tonpad.core.parser.extension.block;
+
+import com.vladsch.flexmark.util.ast.Block;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public abstract class AbstractBlock extends Block {
+}
