+++ src/main/java/org/example/tonpad/core/parser/extension/block/immutable/ImmutableExtension.java
@@ -1,0 +1,26 @@
+package org.example.tonpad.core.parser.extension.block.immutable;
+
+import com.vladsch.flexmark.html.HtmlWriter;
+import com.vladsch.flexmark.html.renderer.NodeRendererContext;
+import org.example.tonpad.core.parser.extension.block.AbstractExtension;
+import org.example.tonpad.core.parser.extension.block.SettingsProvider;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ImmutableExtension extends AbstractExtension<ImmutableBlock> {
+
+    @Override
+    protected SettingsProvider<ImmutableBlock> getSettings() {
+        return new SettingsProvider.SettingsProviderBuilder<>("immutable", ImmutableBlock.class)
+                .build();
+    }
+
+    @Override
+    protected void render(ImmutableBlock node, NodeRendererContext context, HtmlWriter html) {
+        html.attr("contenteditable", "false");
+        html.withAttr().tag("div");
+        context.renderChildren(node);
+        html.tag("/div");
+    }
+
+}
