+++ src/main/java/org/example/tonpad/core/service/MarkdownService.java
@@ -1,0 +1,13 @@
+package org.example.tonpad.core.service;
+
+import com.vladsch.flexmark.util.ast.Document;
+
+public interface MarkdownService {
+
+    Document parseMarkdownFile(String file);
+
+    String renderMarkdownFileToHtml(Document document);
+
+    String convertHtmlToMarkdown(String html);
+
+}
