--- src/main/java/org/example/tonpad/parser/extension/block/AbstractBlock.java
@@ -1,8 +1,0 @@
-package org.example.tonpad.parser.extension.block;
-
-import com.vladsch.flexmark.util.ast.Block;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-public abstract class AbstractBlock extends Block {
-}
