--- src/main/java/org/example/tonpad/service/MarkdownService.java
@@ -1,13 +1,0 @@
-package org.example.tonpad.service;
-
-import com.vladsch.flexmark.util.ast.Document;
-
-public interface MarkdownService {
-
-    Document parseMarkdownFile(String file);
-
-    String renderMarkdownFileToHtml(Document document);
-
-    String convertHtmlToMarkdown(String html);
-
-}
