--- src/main/resources/org/example/tonpad/fxml/file-tree-panel.fxml
@@ -1,90 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.String?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Tooltip?>
-<?import javafx.scene.control.TreeView?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-
-<VBox fx:id="fileTreeVbox" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="184.0" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.tonpad.controllers.FileTreeController">
-   <children>
-      <HBox fx:id="fileTreeToolsHBox" prefHeight="40.0" prefWidth="184.0" spacing="5.0" styleClass="file-tools-panel">
-         <children>
-            <Button fx:id="addNoteButton" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" prefHeight="24.0" prefWidth="24.0" styleClass="round-button" text="N" textAlignment="CENTER">
-               <padding>
-                  <Insets left="3.0" right="3.0" />
-               </padding>
-               <HBox.margin>
-                  <Insets />
-               </HBox.margin>
-               <tooltip>
-                  <Tooltip text="New note">
-                     <styleClass>
-                        <String fx:value="custom-tooltip" />
-                        <String fx:value="tooltip-bottom" />
-                     </styleClass>
-                  </Tooltip>
-               </tooltip>
-            </Button>
-            <Button fx:id="addDirectoryButton" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" prefHeight="24.0" prefWidth="24.0" styleClass="round-button" text="D" textAlignment="CENTER">
-               <padding>
-                  <Insets left="3.0" right="3.0" />
-               </padding>
-               <tooltip>
-                  <Tooltip text="New directory">
-                     <styleClass>
-                        <String fx:value="custom-tooltip" />
-                        <String fx:value="tooltip-bottom" />
-                     </styleClass>
-                  </Tooltip>
-               </tooltip>
-            </Button>
-            <Button fx:id="addCollectionButton" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" prefHeight="24.0" prefWidth="24.0" styleClass="round-button" text="C" textAlignment="CENTER">
-               <padding>
-                  <Insets left="3.0" right="3.0" />
-               </padding>
-               <tooltip>
-                  <Tooltip text="New collection">
-                     <styleClass>
-                        <String fx:value="custom-tooltip" />
-                        <String fx:value="tooltip-bottom" />
-                     </styleClass>
-                  </Tooltip>
-               </tooltip>
-            </Button>
-            <Button fx:id="sortFilesButton" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" prefHeight="24.0" prefWidth="24.0" styleClass="round-button" text="S" textAlignment="CENTER">
-               <padding>
-                  <Insets left="3.0" right="3.0" />
-               </padding>
-               <tooltip>
-                  <Tooltip text="Change sort order">
-                     <styleClass>
-                        <String fx:value="custom-tooltip" />
-                        <String fx:value="tooltip-bottom" />
-                     </styleClass>
-                  </Tooltip>
-               </tooltip>
-            </Button>
-            <Button fx:id="expandFilesButton" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" prefHeight="24.0" prefWidth="24.0" styleClass="round-button" text="E" textAlignment="CENTER">
-               <padding>
-                  <Insets left="3.0" right="3.0" />
-               </padding>
-               <tooltip>
-                  <Tooltip text="Expand all">
-                     <styleClass>
-                        <String fx:value="custom-tooltip" />
-                        <String fx:value="tooltip-bottom" />
-                     </styleClass>
-                  </Tooltip>
-               </tooltip>
-            </Button>
-         </children>
-         <padding>
-            <Insets bottom="5.0" top="5.0" />
-         </padding>
-      </HBox>
-      <TreeView fx:id="fileTreeView" prefHeight="363.0" prefWidth="184.0" styleClass="file-tree-view" VBox.vgrow="ALWAYS" />
-   </children>
-</VBox>
