+++ src/main/resources/ui/css/styles.css
@@ -1,0 +1,113 @@
+/* Стилизация leftPane */
+.side-panel {
+    -fx-alignment: top-center;
+    -fx-background-color: darkGray;
+}
+
+/* Стилизация notePane */
+.note-panel {
+    -fx-background-color: white;
+}
+
+/* Стилизация tabPane */
+.tab {
+    -fx-font-size: 14px;
+}
+
+.tab-pane > .tab-header-area > .tab-header-background {
+    -fx-background-color: gainsboro;
+    -fx-font-size: 14;
+}
+
+.tab-pane > .tab-header-area > .headers-region {
+    -fx-translate-x: 10px;
+}
+
+/* Стилизация treeView */
+.file-tree-view {
+    -fx-background-color: gray;
+}
+
+.file-tree-view .tree-cell {
+    -fx-background-color: gray;
+    -fx-text-fill: white;
+}
+
+.file-tree-view .tree-cell:filled:hover {
+    -fx-background-color: #5a5a5a;
+}
+
+.file-tree-view .tree-cell:filled:selected {
+    -fx-background-color: #404040;
+    -fx-text-fill: white;
+}
+
+.file-tree-view .scroll-bar:vertical {
+    -fx-background-color: gray;
+}
+
+.file-tree-view .scroll-bar:vertical .thumb {
+    -fx-background-color: #5a5a5a;
+}
+
+/* Стилизация fileTreeToolsPane */
+.file-tools-panel {
+    -fx-alignment: top-center;
+    -fx-background-color: dimGray;
+}
+
+/* Стилизация button */
+.round-button {
+    -fx-background-radius: 10%;
+    -fx-border-radius: 10%;
+    -fx-background-color: transparent;
+    -fx-text-fill: white;
+    -fx-cursor: hand;
+    -fx-pref-width: 24px;
+    -fx-pref-height: 24px;
+    -fx-font-size: 14px;
+}
+
+.round-button:hover {
+    -fx-background-color: lightGray;
+    -fx-scale-x: 1.05;
+    -fx-scale-y: 1.05;
+}
+
+.round-button:pressed {
+    -fx-background-color: silver;
+    -fx-scale-x: 0.95;
+    -fx-scale-y: 0.95;
+}
+
+/*Стилизация toolTip*/
+.custom-tooltip {
+    -fx-background-color: #2d2d2d;
+    -fx-text-fill: white;
+    -fx-background-radius: 6px;
+    -fx-border-radius: 6px;
+    -fx-border-color: #555555;
+    -fx-border-width: 1px;
+    -fx-padding: 8px 12px;
+    -fx-font-size: 12px;
+    -fx-font-weight: normal;
+    -fx-effect: dropshadow(gaussian, rgba(0,0,0,0.5), 10, 0.2, 2, 2);
+}
+
+/* Подсказка справа */
+.tooltip-right {
+    -fx-translate-x: 10px;
+    -fx-translate-y: -20px;
+}
+
+/* Подсказка слева */
+.tooltip-left {
+    -fx-translate-x: -30px;
+    -fx-translate-y: -20px;
+}
+
+/* Подсказка снизу */
+.tooltip-bottom {
+    -fx-translate-x: -55px;
+    -fx-translate-y: 15px;
+}
